refactor: get account
