+++ packages/web/src/fixtures/wallet/utility.ts
@@ -1,1 +1,9 @@
 export const isAvailableWallet = () => (window?.ethereum ? true : false)
+
+export const getAccountAddress = () => {
+  const { ethereum } = window
+  if (ethereum) {
+    return ethereum.selectedAddress
+  }
+  return undefined
+}
